+++ src/internal.rs
@@ -4,7 +4,6 @@
 
 use crate::{error::ErrorInner, Error};
 
-
 pub fn deserialize_default<I, O>(src: I) -> Result<O, serde::de::value::Error>
 where
     I: for<'de> serde::de::IntoDeserializer<'de>,
@@ -37,29 +36,53 @@ pub fn map_err_prefix_path<T>(res: Result<T, Error>, prefix: &str) -> Result<T, 
     })
 }
 
+
+macro_rules! get_env_var {
+    ($key:expr, $field:expr) => {
+        match std::env::var($key) {
+            Err(std::env::VarError::NotPresent) => return Ok(None),
+            Err(std::env::VarError::NotUnicode(_)) => {
+                let err = ErrorInner::EnvNotUnicode {
+                    key: $key.into(),
+                    field: $field.into(),
+                };
+                return Err(err.into());
+            }
+            Ok(s) => s,
+        }
+    };
+}
+
 pub fn from_env<'de, T: serde::Deserialize<'de>>(
     key: &str,
     field: &str,
 ) -> Result<Option<T>, Error> {
-    from_env_with(key, field, |de| T::deserialize(de))
+    from_env_with_deserializer(key, field, |de| T::deserialize(de))
 }
 
-pub fn from_env_with<T>(
+pub fn from_env_with_parser<T, E: std::error::Error + Send + Sync + 'static>(
+    key: &str,
+    field: &str,
+    parse: fn(&str) -> Result<T, E>,
+) -> Result<Option<T>, Error> {
+    let v = get_env_var!(key, field);
+    parse(&v)
+        .map(Some)
+        .map_err(|err| {
+            ErrorInner::EnvParseError {
+                field: field.to_owned(),
+                key: key.to_owned(),
+                err: Box::new(err),
+            }.into()
+        })
+}
+
+pub fn from_env_with_deserializer<T>(
     key: &str,
     field: &str,
     deserialize: fn(crate::env::Deserializer) -> Result<T, crate::env::DeError>,
 ) -> Result<Option<T>, Error> {
-    let s = match std::env::var(key) {
-        Err(std::env::VarError::NotPresent) => return Ok(None),
-        Err(std::env::VarError::NotUnicode(_)) => {
-            let err = ErrorInner::EnvNotUnicode {
-                key: key.into(),
-                field: field.into(),
-            };
-            return Err(err.into());
-        }
-        Ok(s) => s,
-    };
+    let s = get_env_var!(key, field);
 
     match deserialize(crate::env::Deserializer::new(s)) {
         Ok(v) => Ok(Some(v)),